requirements.txt
pygame==2.5.2
pyyaml==6.0.1
pillow==10.1.0
numpy==1.26.2
cherrypy==18.8.0
